Extract pre-checkin modals to ModalManager
